Move types from NavigationBar to type file
